[move] DTOs to relevant folder
