+++ src/main/java/com/example/banking_app/dto/Withdraw.java
@@ -1,4 +1,4 @@
-package com.example.banking_app.model;
+package com.example.banking_app.dto;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -6,10 +6,10 @@ import java.math.BigDecimal;
 
 @Getter
 @Setter
-public class Deposit {
+public class Withdraw {
 
     private BigDecimal amount;
 
-    public Deposit() {
+    public Withdraw() {
     }
 }
